CB-2245 save script remove exception when file doesn't exist

+++ server/bundles/io.cloudbeaver.model/src/io/cloudbeaver/model/rm/local/LocalResourceController.java
@@ -336,6 +336,9 @@ public class LocalResourceController implements RMController {
                     data.length);
         }
         Path targetPath = getTargetPath(projectId, resourcePath);
+        if (!Files.exists(targetPath.getParent())) {
+            throw new DBException("Parent folder '" + targetPath.getParent().getFileName() + "' doesn't exist");
+        }
         try {
             Files.write(targetPath, data);
         } catch (IOException e) {
@@ -370,9 +373,6 @@ public class LocalResourceController implements RMController {
             Path targetPath = projectPath.resolve(resourcePath).normalize();
             if (!targetPath.startsWith(projectPath)) {
                 throw new DBException("Invalid resource path");
-            }
-            if (!Files.exists(targetPath.getParent())) {
-                throw new DBException("Parent folder '" + targetPath.getParent().getFileName() + "' doesn't exist");
             }
             return WebAppUtils.getWebApplication().getHomeDirectory().relativize(targetPath);
         } catch (InvalidPathException e) {
